+++ src/app/[orgId]/settings/(private)/access/approvals/page.tsx
@@ -1,8 +1,9 @@
-import { ApprovalFeed, type ApprovalItem } from "@app/components/ApprovalFeed";
+import { ApprovalFeed } from "@app/components/ApprovalFeed";
 import SettingsSectionTitle from "@app/components/SettingsSectionTitle";
 import { internal } from "@app/lib/api";
 import { authCookieHeader } from "@app/lib/api/cookies";
 import { getCachedOrg } from "@app/lib/api/getCachedOrg";
+import type { ApprovalItem } from "@app/lib/queries";
 import OrgProvider from "@app/providers/OrgProvider";
 import type { GetOrgResponse } from "@server/routers/org";
 import type { AxiosResponse } from "axios";
